Vtlanglois/issue20 (#22)
