+++ src/pages/about/About.tsx
@@ -1,7 +1,7 @@
 import { faUser } from "@fortawesome/free-solid-svg-icons";
 import { Container, Fade, Row } from "reactstrap";
-import TabHeader from "../../components/ExperienceItem/TabHeader";
-import WorkAccordian from "../../components/WorkAccordian/WorkAccordian";
+import TabHeader from "../../components/about/ExperienceItem/TabHeader";
+import WorkAccordian from "../../components/about/WorkAccordian/WorkAccordian";
 import "../../App.css";
 const About = () => {
   return (
